[REFACTOR] auth 컴포넌트 구조 분리
